[feat] 4단계 에니메이션 구현
